[SPARK-42078][PYTHON] Migrate errors thrown by JVM into `PySparkException`
### What changes were proposed in this pull request?

This PR proposes to migrate errors thrown by JVM into `PySparkException`.

Here is the full list of errors suggesting migration:
- CapturedException (The parent class for every errors thrown by JVM)
  - AnalysisException
  - ParseException
  - IllegalArgumentException
  - StreamingQueryException
  - QueryExecutionException
  - PythonException
  - UnknownException
  - SparkUpgradeException

In short, this PR proposes to migrate all exceptions from `pyspark/sql/utils.py` into `pyspark/errors/exceptions.py` and fix them to inherit `PySparkException`.

### Why are the changes needed?

Because the errors thrown by JVM side (e.g. `AnalysisException`, `ParseException `, `IllegalArgumentException ` ... ) are managed by `pyspark/sql/utils.py` which is not very correct place to handle the PySpark errors so far.

Now we have separate error package `pyspark.errors` which is introduced from https://github.com/apache/spark/pull/39387, we should centralize whole errors generated by PySpark side into single management point.

So, here I suggest that all errors occurring in the JVM inherit PySparkException so that errors can be managed consistently.

### Does this PR introduce _any_ user-facing change?

No, this PR suggests to improve an internal error handling structure without any user-facing change.

### How was this patch tested?

The existing CI should pass.

Closes #39591 from itholic/SPARK-42078.

Authored-by: itholic <haejoon.lee@databricks.com>
