+++ python/docs/source/reference/pyspark.errors.rst
@@ -20,6 +20,28 @@
 Errors
 ======
 
+Classes
+-------
+
+.. currentmodule:: pyspark.errors
+
+.. autosummary::
+    :toctree: api/
+
+    PySparkException
+    AnalysisException
+    ParseException
+    IllegalArgumentException
+    StreamingQueryException
+    QueryExecutionException
+    PythonException
+    UnknownException
+    SparkUpgradeException
+
+
+Methods
+-------
+
 .. currentmodule:: pyspark.errors
 
 .. autosummary::
@@ -27,3 +49,4 @@ Errors
 
     PySparkException.getErrorClass
     PySparkException.getMessageParameters
+    PySparkException.getSqlState
