+++ python/pyspark/errors/__init__.py
@@ -18,9 +18,27 @@
 """
 PySpark exceptions.
 """
-from pyspark.errors.exceptions import PySparkException
+from pyspark.errors.exceptions import (  # noqa: F401
+    PySparkException,
+    AnalysisException,
+    ParseException,
+    IllegalArgumentException,
+    StreamingQueryException,
+    QueryExecutionException,
+    PythonException,
+    UnknownException,
+    SparkUpgradeException,
+)
 
 
 __all__ = [
     "PySparkException",
+    "AnalysisException",
+    "ParseException",
+    "IllegalArgumentException",
+    "StreamingQueryException",
+    "QueryExecutionException",
+    "PythonException",
+    "UnknownException",
+    "SparkUpgradeException",
 ]
