+++ python/pyspark/sql/tests/pandas/test_pandas_udf_scalar.py
@@ -49,7 +49,7 @@ from pyspark.sql.types import (
     DateType,
     BinaryType,
 )
-from pyspark.sql.utils import AnalysisException
+from pyspark.errors import AnalysisException
 from pyspark.testing.sqlutils import (
     ReusedSQLTestCase,
     test_compiled,
