+++ python/pyspark/sql/tests/streaming/test_streaming.py
@@ -236,7 +236,7 @@ class StreamingTests(ReusedSQLTestCase):
             sq.stop()
 
         from pyspark.sql.functions import col, udf
-        from pyspark.sql.utils import StreamingQueryException
+        from pyspark.errors import StreamingQueryException
 
         bad_udf = udf(lambda x: 1 / 0)
         sq = (
@@ -479,7 +479,7 @@ class StreamingTests(ReusedSQLTestCase):
         self.assertEqual(len(tester.close_events()), 0)
 
     def test_streaming_foreach_with_process_throwing_error(self):
-        from pyspark.sql.utils import StreamingQueryException
+        from pyspark.errors import StreamingQueryException
 
         tester = self.ForeachWriterTester(self.spark)
 
@@ -595,7 +595,7 @@ class StreamingTests(ReusedSQLTestCase):
                 q.stop()
 
     def test_streaming_foreachBatch_propagates_python_errors(self):
-        from pyspark.sql.utils import StreamingQueryException
+        from pyspark.errors import StreamingQueryException
 
         q = None
